runcommand.sh get_all_x11_modes speedup, translates bash logic to awk

+++ scriptmodules/ports/kodi.sh
@@ -21,7 +21,8 @@ function _update_hook_kodi() {
 }
 
 function depends_kodi() {
-    if isPlatform "rpi"; then
+    # raspbian
+    if [[ "$__os_id" = "Raspbian" ]] && isPlatform "rpi"; then
         if [[ "$md_mode" == "install" ]]; then
             # remove old repository
             rm -f /etc/apt/sources.list.d/mene.list
@@ -31,8 +32,17 @@ function depends_kodi() {
             rm -f /etc/apt/sources.list.d/pipplware.list
             apt-key del 4096R/BAA567BB >/dev/null
         fi
-    elif isPlatform "x86" && [[ "$md_mode" == "install" ]]; then
-        apt-add-repository -y ppa:team-xbmc/ppa
+    # ubuntu
+    elif [[ -n "$__os_ubuntu_ver" ]] && isPlatform "x86"; then
+        if [[ "$md_mode" == "install" ]]; then
+            apt-add-repository -y ppa:team-xbmc/ppa
+        else
+            apt-add-repository --remove -y ppa:team-xbmc/ppa
+        fi
+    # others
+    else
+        md_ret_errors+=("Sorry, but kodi is not installable for your OS/Platform via RetroPie-Setup")
+        return 1
     fi
 
     # required for reboot/shutdown options. Don't try and remove if removing dependencies
@@ -44,7 +54,19 @@ function depends_kodi() {
 function install_bin_kodi() {
     # force aptInstall to get a fresh list before installing
     __apt_update=0
-    aptInstall kodi kodi-peripheral-joystick kodi-inputstream-adaptive kodi-inputstream-rtmp kodi-vfs-libarchive kodi-vfs-sftp kodi-vfs-nfs
+
+    # not all the kodi packages may be available depending on repository
+    # so we will check and install what's available
+    local all_pkgs=(kodi kodi-peripheral-joystick kodi-inputstream-adaptive kodi-inputstream-rtmp kodi-vfs-libarchive kodi-vfs-sftp kodi-vfs-nfs)
+    local avail_pkgs=()
+    local pkg
+    for pkg in "${all_pkgs[@]}"; do
+        # check if the package is available - we use "madison" rather than "show"
+        # as madison won't show referenced virtual packages which we don't want
+        local ret=$(apt-cache madison "$pkg" 2>/dev/null)
+        [[ -n "$ret" ]] && avail_pkgs+=("$pkg")
+    done
+    aptInstall "${avail_pkgs[@]}"
 }
 
 function remove_kodi() {
